integrate glean2pysa with model generator
Summary:
This change introduces a model generator which combines 3 disparate sources of data: Zoncolan, Pyre and Glean, in order to map user-controlled WWW endpoints into the Python functions that are sources of user controlled taint, and generat\
e appropriate models.

The basic idea is that Glean gives us a set of hack-to-python thrift connections (without regard of which parameters are user controlled). Zoncolan keeps track of which thrift endpoints are user controlled. Pyre gives us Python AST knowl\
edge which allows us to map the Glean + Zoncolan output to a python function, and generate the model.

In order to run with a target, you need to run 3 steps - generate the glean data,  generate the fake buck target, and run the model generator.

Reviewed By: fahndrich

Differential Revision: D17889496

fbshipit-source-id: 1f286609c595d41dbb0bddca8751a37c640e50eb