Introducing TypeEnvironment
Summary:
Context
======

The current API for all checks, including type checking, looks roughly like this (simplified from `Analysis.Check.Signature.run`:

```
val run: Source.t ->  Error.t list
```

Note how this API is module-based: one can't use it to pick several functions within a module and only recheck those. This is not desirable for type checking since we'd want the check granularity to be finer, but there are other checks, such as inference and deobfuscation, that do rely on module-level info. Therefore, it is not possible to just nuke this API and replace it with another one for the sake of fine-grained typechecking :(.

One idea here is to take `TypeCheck` out of the `Check` infrastructure and let it do its own thing, but my estimation is that the approach is going to cause a much larger-scale change that I am not feeling too comfortable to deal with at the moment.

Alternatively, since we current don't care about incremental runs for any checks other than typecheck, the basic approach I ended up taking is to (1) keep the current module-based API for full checks, and (2) have an additional function-based API for incremental invalidation and re-type-check. The tricky part of this design is how we can connect (1) and (2) for typecheck -- we'd like to be able to associate the errors produced by module-level checks to individual functions, so when we run an incremental check later we can selectively perform function-level invalidation and rechecks.

NOTE: This association cannot be derived just by looking at each error: Although `BaseError.signature` exists (it really shouldn't :(), it only tells us the function from which the error gets emitted, not the function from which the error computation is scheduled. Until we are able to schedule all function checks independently, those two are *not* going to be the same thing.

We can enforce each check to establish this association by changing their module-based APIs to be:

```
val run: Source.t -> ((Error.t list) Reference.Map.t) list
```

And the expectation is that in the result list, errors emitted by each scheduling unit (i.e. `$toplevel` function and its direct descendants) are cleanly separated. If we store the results somewhere, incremental checks in later stage can leverage it to perform finer grained invalidation correctly.

Implementation
============

My implementation actually changes `Analysis.Check.Signature.run` to be stateful:

```
val run: Source.t -> TypeEnvironment.t -> unit
```

where `TypeEnvironment.t` is essentially `(Error.t list) Reference.Map.t` except it's in shared memory. This is just me trying to simplify the "store the results somewhere" part. Conceptually it should not be too different from the previous signature where a `((Error.t list) Reference.Map.t) list` is returned. In the future we might also be interested in moving other shared memory tables into `TypeEnvironment`, but for now let's get started with just errors.

To facilitate code review, this diff only introduces the `TypeEnvironment` APIs. Connecting the environment to prod is done in the stacked diff above (D18301094).

Reviewed By: mrkmndz

Differential Revision: D18301094

fbshipit-source-id: bb22e0cf71235d54508426b40cf6b6fd9bea4ac9