Keep track of the name of nesting define for all nested defines
Summary:
No matter what we do, this is very useful for fine-grained dependencies. A nested define needs to know what other define(s) it is nested inside, in order to:

- Register type variables from the outer function (nested function should be able to "see" typevars from their nesters).
- Potentially try to access the bindings of the outer functions, to figure out which variables are captured.
- Figure out which typecheck unit it belongs to, if we still want to head towards the direction of separating tcu vs. function (unlikely)

The info is not hard to get since it is completely syntactical. This diff modifies the parser to have it "inject" the name of the nester directly into the AST.  Relevant changes can be found in `statement.ml` and `generator.mly`.

Reviewed By: dkgi

Differential Revision: D18375842

fbshipit-source-id: af7267680bdf1494b56fa50ff1dcb94c96fd028b