Break infinite loop when solving constraints for callable parameters
Summary:
We need to break the logic loop that was causing stack overflows in any class that has a Callable type as the first method parameter:

```
class Foo:
    a: int
    def derp(herp: Callable[..., Any]) -> None: ...
```

The logical loop:

1. Look up any attribute in `Foo`.
2. Triggers population of all attributes in Foo
3. To populate the attribute `derp`, we solve the first parameter against `Foo`.
4. Comparing a callable with a class means we need to look up the `__call__` attribute in the class
5. [back to step 2]

While the example above doesn't make a ton of sense in Python, it at least demonstrates how a single parameter annotated in this way can create stack overflows regardless of what attribute you are looking up. The following is legitimate, type-safe Python that demonstrates how the nature of solving for callable parameters is naturally recursive:

```
class Foo:
    def __call__(self: Callable[[int], str], x: int) -> str: ...
```

This is valid and should type check. If you changed the annotation of `x` to something else, this should error.

It's not actually possible, theoretically, to disentangle attribute lookup from needing to look back inside the parent class for more information about its attributes. This boils down to the same kind of recursion that happens when we try to solve whether a class implements a generic protocol. The key difference here is that we cannot represent Callable as a protocol without support for classes that are generic over TParams.

Until that is added, we can break this cycle with a simple parallel to ProtocolAssumptions that we'll call CallableAssumptions, so we can assume compliance and verify the assumption in the second loop around.

Reviewed By: mrkmndz

Differential Revision: D18384215

fbshipit-source-id: 562c49eef083e106e213f52d52efc260359efbfe