Temporarily fix access of arguments fields from shim
Summary:
This is a hacky patch for now, but the underlying problem is: we shouldn't be manipulating the arguments namespace object to modify fields or add additional ones. This is how switch_root and other preprocessing of arguments was done leading up until now, and executing the shim reveals problems because we care about setting up logging, but just passing in the same arguments object can break because the necessary argument preprocessing steps weren't all run.

A more robust fix is coming up, where I'll set up a concrete class for arguments that will verify the appropriate ones exist, and explicitly define any additional fields we care to add. Ideally, any preprocessing steps can be centralized in this arguments module rather than being scattered in client/__init__ or client/pyre, so we can be sure when we pass arguments objects around we know exactly what's been done to it and what fields it contains.

Reviewed By: sinancepel

Differential Revision: D18402900

fbshipit-source-id: b94b94666b4b769b15a49c8b02e4bb636fea0cf6