simulate the proper statement for `with` and `async with`
Summary:
Our preamble construction for `with` and `async with` was broken in the case where we didn't use the `with x as y` form. This caused us to not register some expressions as being awaited.

Also added some more detailed printing to the cfg tests as they were impossible to debug beforehand.

Reviewed By: dkgi

Differential Revision: D18359217

fbshipit-source-id: fb4187aa6f0ea70e83e843438360528105f944fa