Poll for completion rather than calling incremental in incremental_test
Summary:
The current timing model for the incremental_test is to do
```
update * n
time {
  pyre incremental
}
validate incremental output
```
This does not give us visibility into the sub-phase times and is also inaccurate, since the updates themselves take some time in which the underlying update job is already running on the server.

This diff converts it to
```
for n updates:
    update
    poll until get new result {
        pyre profile --output=incremental_updates
    }
pyre incremental # just for the already computed errors
validate incremental output
```
This gives us all of the detail we want, and is much more robust.

Reviewed By: grievejia

Differential Revision: D18381533

fbshipit-source-id: bf7dbed23d0f5eec38bf9cac3d227811952a9be8