Add a method to rebuild the scratch analysis directory.
Summary:
Problem: When there is a rebase in fbcode, `pyre incremental` doesn't update correctly on new or deleted files.

Solution: Add a method that asks buck to build the targets and uses the new buck output directory to update the symbolic links and the inverse symbolic links dictionary. This will now notify the server of files added or deleted during the rebuild.

Currently, this calls `rebuild` when the number of updated tracked files exceeds a threshold or the number of new or deleted files exceeds a threshold.

Reviewed By: sinancepel

Differential Revision: D18279965

fbshipit-source-id: b584cd645373e8a869a7ceda97853395b20bc5d1