Move global parser options to base command class
Summary:
(Attempting follow-up from D18405673)

Goal: we should never be passing around, appending to, reading from the arguments namespace object. Currently a lot of client setup is brittle because fields are assumed to exist, or because we rely on preprocessing of arguments like switch_root, translate_arguments, find_log_directory in order for the set of arguments to contain all necessary and correct elements.

There isn't anything guaranteeing that certain preprocessing calls have happened, or that certain fields exist on the arguments object.

Trying to formalize this by moving the definition of each argument into its command class, and ensuring the only thing reading from the raw arguments namespace is the initialization of each command class

Still TODO:
- Refactor configuration creation to happen in the base command initialization, as it requires reading from arguments.
- Refactor analysis directory creating, as currently this requires the input of configuration and arguments, and command construction currently requires input of analysis directory and arguments. We need to move this into the command object, have subcommands override it where needed, and also be able to return it into init so we can clean it up in the finally block.
- Refactor all preprocessing operations like switch root, translate arguments, find log directory, to rely only on arguments all commands will have, and to occur during init of base command.
- Parse all arguments in constructors of commands (set defaults for arguments that are not given on command line)
- Either refactor shim logic to go through a command, or special case it carefully to look through arguments...? Or get rid of logging there. Need to think more about this :P
- Clean up all locations that look up fields in arguments, where fields don't always exist because they are specific to a particular subcommand.

Reviewed By: sinancepel

Differential Revision: D18429300

fbshipit-source-id: 3d08411517494e262d3f519920a576411ab18723