removing duplicated check of function body
Summary:
In a early version, type checking function body was done recursively by
the `typ` predicate. Now that we have the `well_formed_function` property,
it is redundant. Removing it and using only the `well_formed` status of the
typing context.
Now the correctness of typing function calls depends on the correctnesso of
the context. Locally, we trust the context information.

Reviewed By: dkgi

Differential Revision: D18449119

fbshipit-source-id: 89c746ffc9c5289911f1e946d665e3798eb1de07