Refactor analysis directory creation
Summary:
Refactor analysis directory creating, as currently this requires the input of configuration and arguments, and command construction currently requires input of analysis directory and arguments. We need to move this into the command object, have subcommands override it where needed, and also be able to clean it up in the finally block.

Also consolidated the options build/isolate into subcommand overrides rather than checks on arguments.command against a list of commands.

Reviewed By: dkgi

Differential Revision: D18438670

fbshipit-source-id: d62b7ce5bcd1eeecd6cd33d89ceddbf0d888884f