Dependency-track function bodies
Summary:
NOTE: This is pretty much a re-submission of D18333528, except that we are now dealing with functions directly, instead of messing with typecheck units.

Goal
====

We want 2 useful APIs for function-level typecheck:
- `get_defines_in_module`: Given a module name, return the name of all functions it contains
- `get_define_body`: Given a function name, return the body of the function, if it exists.

We also want the result of these two APIs to be in shared memory with dependency tracked.

Why
======

Let's start with `get_define_body`: this is just a useful operation to have in general, even for today: Except for the subtleties I'm going to describe below, Sinan's `defines()` query (D18339919) essentially boils down to this. I also need this API to grab the corresponding signature from a define name -- combined with D18375842, we should be able to get the signatures of all other defines that nest a given define, enabling us to propagate type variables across nested functions. Type variables from nesting functions is one of the two blocker of function-level typechecking (the other blocker being captured variables).

Note that this API does have one subtle semantics implication: it assumes that a name uniquely defines a function. Although the majority of our codebase seems to make this assumption (e.g. `ResolutionSharedMemory` is keyed by function name), it is not always guaranteed: `overload`-ed function stubs share the same name with the implementation, and developers may also (intentionally or unintentionally) write functions with duplicated name. Having spent quite some effort in fighting with this assumption in our codebase with no success, I'll just succumb to the status quo and try to enforce it instead:
- For overloaded functions, only the implementation counts for the function body. Overloaded stubs are silently excluded.
- For functions with duplicated names, we only keep the body of the last occurrence. This is consistent with how we currently handle duplicated class definitions.

-----------------------------

For `get_defines_in_module`, at first glance it seems that the information could be obtained simply by invoking `Preprocessing.defines` on the corresponding module instead. However, `Preprocessing.defines` does not collect every defines: functions and classes that are nested inside `if`, `while`, `try`, etc. would be ignored. As there are too many users the relies on `Preprocessing.defines` and I don't have the full contexts on each uses, it's easier for me to just write my own collector as opposed to patch the existing one.

------------------------------

The reason why we want the info in shared memory and dependency tracked is to support function-level incremental check:

- Previously, whenever a source changes, we simply assume all defines in the source changes accordingly and recheck them all in the typechecking phase. This is clearly no longer a good strategy -- with dependency triggers on `get_define_body`, we are able to avoid re-check functions whose bodies do not change (provided that they are not triggered by anything else).
- Previously, whenever a source changes, we re-compute all defines it has in the typechecking phase by running `Preprocessing.defines` on the entire source. This is no longer a valid solution for fine-grained incremental checks: without the luxury to simply nuke all info for the old module entirely, it must start to maintain the set of functions a module contains as it can change in different ways when a module changes. For example, if a function `foo()` gets removed from the module, invoking `Preprocessing.defines` on the updated module will not include `foo` at all -- to correctly invalidate shared memory entry for `foo`, we'll need to somehow keep track of the set of all functions in the original module as well as the set of all functions in the newly updated module. This is exactly what motivate us to build dependency-tracked table in the first place!

How
============

We add both APIs to `UnannotatedGlobalEnvironment`, because this is naturally how other "fan-out" entities (i.e. class definitions and unannotated globals) are tracked in the existing system. The change itself is quite mechanical as a result: I simply follow the existing code pattern for class definitions and unannotated globals.

This diff only exposes the API, and itdoes not add and use for it. The plan is to eventually migrate the type checker to grab function bodies solely via the introduced APIs -- anything that's not collected here will not get type checked.

NOTE: In the final revision, `get_defines_in_module` gets renamed to `all_defines_in_module`, and this API is *not* dependency tracked. As a result, we should never use it in the context of incremental checks.

Reviewed By: mrkmndz

Differential Revision: D18386508

fbshipit-source-id: 03886f1b051818078ae896a1a0a534a5446bf67b