Proper scope handling I: bindings and non-nested scopes
Summary:
Context
======

To be able to type check each function individually, one need to correctly propagate local variables in the and their annotations of a nesting function to all inner functions it nested. For example,

```
def foo(x: int) -> int:
  def bar(y: int) -> int:
    return x + y  # <-- Need to propagate the binding of "x: int" from foo()
  return bar(42)
```

To correctly type check `foo.bar()` without type checking `foo()`, we'll need to somehow tell `foo.bar()` that:
- A name `x` is defined in `foo()`, so we should avoid emitting an "`x` not defined" error in `foo.bar()`.
- `x` is annotated with type `int`, so we don't lose information when trying to infer the type of expressions like `x + y`  in `foo.bar()`.

Generally, handling captured variables requires us to correctly compute the following info:
- Given an outer function (e.g. `foo()`), what is the set of *bindings* (i.e. name and associated type annotation) it contains. As mentioned before, we need to make some of these bindings visible to the inner functions to grant them type-checking independency.
- Given an inner function (e.g. `foo.bar()`), what is the set of bindings it accesses that are captured (i.e. not defined in the same scope) from its nesting functions? Not all bindings in the outer functions are necessary for the type checking of the inner function -- if a name is never accessed in the inner function, we can filter the corresponding bindings out from the outer function to avoid pulling unnecessary data into the local annotation maps.
- If there are multiple levels of nesting, how would the binding sets of different nested levels interact with each other? This is further complicated by Python's `global` and `nonlocal` statement, which subtly changes the way those interactions work.

Unfortunately, the existing Pyre codebase offers no easy way to compute any of those info. The closest thing we get is `Preprocessing.qualify`, but the way names get mangled in the qualification pass is rather arbitrary and not suitable to be relied upon for captured variable detection.

Let's try to address the issue with this series of "proper scope handling" diffs. We'll gradually build up the toolset we need to collect binding and resolve accesses from the first principle. The short-term end goal for this diff stack (est. a handful of diffs) is to have a new preprocessing pass that, for each function, inject the set of captured bindings to the function's `Define` node. In the long run, we might consider rewiring the qualification pass to base on this toolset as well.

The most helpful Python document I found when implementing scopes can be found [here](https://docs.python.org/3/reference/executionmodel.html).

Implementation
============

This diff implements the notion of *binding* and (non-nesting) *scope*.

Binding
--

A `Binding.t` represent how a name gets introduced. For our purpose, we are interested in the name itself, the kind of introduction (e.g. assignment, parameter, import, etc.), the location where the introduction happens (potentially useful for error reporting), and the associated type annotations for the name.

The first notable thing about bindings is that they are mostly *lexical*. In other words, to determine the set of bindings at a give location, we need to look at nothing other than the AST. The *value* stored in the bindings may change at runtime, but the bindings themselves shouldn't. For example,

```
x = 42
def foo(y):
  x = y
  while bar(x):
    x = x + 1
  return x
...
```

Here the value of `x` may depends on how `foo()` is invoked and what `bar()` returns at runtime, but all references to `x` in `foo()` always point to the `x` bound at line 3, not the one bound at line 1.

The lexical nature of bindings means that we are able to handle scopes purely in the parsing/preprocessing step.

NOTE: There are several special cases in Python where bindings are dynamically determined (`exec()`, `eval()` and `class`). In general, this kind of dynamic bindings cannot be fully supported by a static analyzer. So to make our jobs simpler, we'll treat bindings as exclusively lexical for now, and worry about dynamic bindings in the future.

The second notable thing about bindings is that they occur at the granularity of code-block:  all uses of the name within a code block are treated as references to the current block. In Python, only modules, function bodies, and class definitions count as code block. AST entities like `if` body, `while` body, `try` body, etc. do *not* make new code block.This leads to some quite unintuitive behaviors:

```
def example0(flag):
  if flag:
    print(x)  # Refers to the x bound at the else branch (flag=True -> UnboundLocalError)
  else:
    x = 1

def example1():
  def inner():
    print(x)  # Refers to the x bound at the line below (*no* runtime error)
  x = 1
  inner()
```

The per-block nature of bindings means that we are able to perform binding collections in a flow-insensitive manner.

The third notable thing about bindings is the various way they get introduced. One might expect that parameters and assignments would introduce bindings, which is correct. But one might also expect that `for` loops would not introduce bindings, which is incorrect: Python would leak the name of the for loop target to the containing block:

```
s = 0
for i in range(10):
  s += i
print(i)  # <-- Refers to the i bound at line 2
```

Up until Python 3.7, the following constructs can bind names:
- Function parameters
- Import statements
- Class and function definitions (these bind the class or function name in the defining block)
- Assignment, for loop header, with statement or except clause.
- Lambda parameters, comprehension targets

Python 3.8 adds walrus operator to the picture, which would have a big impact on binding introductions (expressions can now introduce bindings to the enclosing scope, which is not possible in the past). We'll leave walrus operator handling as a TODO for now.

Scope
--

For our purpose, a scope is just a collection of bindings keyed by the binding names.

Within a scope, we can also declare a name as global or nonlocal using the `global`/`nonlocal` statement. If a name is declared as `global` or `nonlocal`, we know that the name, if exists, must come from other scopes -- therefore we would exclude those names from the set of local bindings unconditionally.

`global`/`nonlocal` applications follow the same granularity as bindings: If the `global`/`nonlocal` statement occurs within a block, it applies to all uses of the name specified in any statement of the block. This means that the following two code examples are equivalent:

```
x = 1
def example0(flag: bool):
  global x
  if flag:
    x = 1  # Refers to the global x
  else:
    pass

def example1(flag: bool):
  if flag:
    x = 1  # Refers to the global x as well
  else:
    global x
```

Due to the intricate nature of `global`/`nonlocal` handling, in our `Scope` API we do not allow clients to create scopes out of a single statement -- they must be created from an entire `Define`, an entire `Source`, or some specific expressions (since expressions can never introduce `global`/`nonlocal`).

In this diff, we do not handle scope nestings at all -- `Scope` constructors will not look into nested defines/expressions. Nestings will be handled separately in the next diff.

Reviewed By: dkgi

Differential Revision: D18441819

fbshipit-source-id: 6ee5a693f6a174458bdd4bfe0eb7bf88e57b0b7f