Move nesting define propagation logic after qualification
Summary:
We were not qualifying the `nesting_define` field for nested functions correctly. Consider this:
```
def foo():
   def foo():
      pass
```
Before qualification, the `nesting_define` field of the inner `foo()` function is set to `foo`. After qualification, however, the `nesting_define` field will be set to the *same name* as the inner `foo()` function. The downstream analyses would then consider that the inner `foo()` is nested in itself, creating an infinite loop in the type checker due to my recent change in D18400709.

I've tried hard, but the sad reality is that `Preprocessing.qualify` is just too broken to qualify the `nesting_define` field correctly. The crux of the problem is that `nesting_define` needs to be qualified in the *outer* scope of the function, while we only consider the *innermost* scope when qualifying a statement. Properly fixing this issue requires a big restructure of the qualification pass, which I don't think is worth the effort since there are so many problems that we'll need a more fundamental rewrite of it eventually anyway.

Instead, the workaround I propose here is to move the pass that calculate the `nesting_define` field out of the parser and into a standalone preprocessing pass, so that nesting defines can be computed *after* qualification has already taken place.

Reviewed By: sinancepel

Differential Revision: D18481712

fbshipit-source-id: 4d140bb7b3872999209ed108e1cc51cccc498a72