Do not reuse local variable names when qualifying function names
Summary:
We were not qualifying nested functions correctly. This function:

```
# qualifier.py
def foo():
  def foo():
    def foo():
      pass
```
gets qualified into this:
```
def qualifier.foo():
  def $local?qualifier?foo$foo():
    def $local?qualifier?foo$foo():
      pass
```
The innermost `foo` and the second innermost `foo` are qualified to the same name, which confuses downstream analyses because they were all assuming that function names uniquely identifies the function body.

Turns out that the root cause is that when qualifying function names, we count function names as locals and also try to reuse existing local names as much as possible. As a result when qualifying the innermost `foo()`, we see the outer `foo()` in the collection of locals and reuse the name for it :(

Reviewed By: sinancepel

Differential Revision: D18471102

fbshipit-source-id: 199f3b0d74ba7f8c6c36fa48ab98795a5a0009e4