fix inference crash for x[0] += y[1]
Summary:
We were calling name_to_reference_exn for non-simple names.

I took this as an opportunity to purge most uses of name_to_reference_exn, as introducing a helper to check whether a name could be converted to a reference seemed to make the code more readable and save repetitive conversions in the body of backward.

Reviewed By: MaggieMoss

Differential Revision: D18487517

fbshipit-source-id: 42663de84ff0e0f6a6c41359c07291764a3db974