Handle new or deleted file using buck query.
Summary:
Problem: When the number of files added or deleted was below a threshold, we didn't rebuild the scratch directory.

This diff adds or deletes a symbolic link and updates the symbolic link map using a buck query.

It also finds relative paths only for files covered by the stored set of targets. Otherwise, we could add files that are outside of our project.

Will check in upcoming diffs if the new files are special files like TARGETS, etc.

Reviewed By: sinancepel

Differential Revision: D18459760

fbshipit-source-id: e7355d77173a0c1cd714ca1d3e001036ff35fadd