Converting all existing Analysis.Environment tables to be lazily filled (in fine grained mode)
Summary:
It actually turns out that converting the existing stock of environment tables to be lazily filled was relatively straightforward.  The update behavior is almost exactly the same, except for that we don't produce values for additions ahead of time.  Triggered dependencies are still processed, but we can wait to compute the values for new keys until they are actually asked for.
In theory this could have us doing less work.

The main purpose of this is not for perf, but rather to have this as a tool to build managed caches.

Theoretically this diff introduces some extra work in the case of looking up what would previously have been a missing key in legacy mode, as this causes you to have to do a fresh computation which will (hopefully) return nothing.  However, it turns out that this doesn't actually impact perf in any significant way, and is still correct behavior, so doesn't seem worth struggling to have get's behavior depend on configuration as well.

Note:
This kind of messes with our sub-timing profiling scheme, as additions will now be counted against type check time, rather than the actual phase it corresponds to.  This is less than ideal, but I believe, worth it.

Reviewed By: dkgi

Differential Revision: D18471442

fbshipit-source-id: 0ddcb1555bf436c3e482ab6fbcd50be8b2a1fffe