Making legacy mode/full check lazy environment build as well
Summary:
Now that we have set up the environment get to be able to populate the field on demand, we can disable legacy mode/full check from populating everything ahead of time.  By doing this, we remove the need for access to the added keys at any point, and only need the set of potentially modified keys in legacy mode.

This accomplishes a long standing goal: lazy environment building.

Instead of calculating every conceivable property we'd ever possibly need ahead of time, we instead calculate it whenever we need it, in a manner that's cached and shared across workers.  This means that we avoid doing a lot of work that is unnecessary.

The hard work that enabled this win was done in all of the restructuring of environment building to be "keyed on triggers", i.e. being able to produce a key given only that key and the previous environments.

Reviewed By: grievejia

Differential Revision: D18487053

fbshipit-source-id: 9d9098e33bbf5aae61d45d72e9daf661fc92ffd1