Proper scope handling II: nested scopes and lookup
Summary:
Context
======

See D18441819

Implementation
============

This diff implements scope stacking (exposed via the `ScopeStack.create` and `ScopeStack.extend`) and binding lookup (exposed via `ScopeStack.lookup` function)

If we get the non-nested scopes right, nested scopes are surprisingly straightforward to implement:
- A `ScopeStack` is, as the name suggests, a stack of scope :)
  One would initialize such a stack with `ScopeStack.create` (which takes `Source.t` since the module-level scope is always on the bottom of the stack), and stack other scopes on top of it with `ScopeStack.extend`.
- Looking up `ScopeStack` simply means walking through the stack from top to bottom, looking up each scope along the way, and return the first binding we find if there's a hit.

NOTE: Given a particular function, the right stack for it needs to be constructed manually by walking down the AST from the client side for now -- In this diff, I did not include an API that takes a source and return a list of scope stacks for each function. Such a powerful API might be useful, but what I found hard is to figure out a way to associate scope stacks to their corresponding functions -- remember that before `UnannotatedGloalEnvironment.update` takes place, we can't key functions by their names. For now, let's leave those design questions to the future if necessary.

What complicates the picture, however, is the existence of `global` and `nonlocal` declarations. From the doc:

> If the global statement occurs within a block, all uses of the name specified in the statement refer to the binding of that name in the top-level namespace.

This means that if a variable is declared as `global` in a given scope, we should skip all the local scopes and immediately jump to the outermost scope of the module and perform the search there. Also,

> The nonlocal statement causes corresponding names to refer to previously bound variables in the nearest enclosing function scope.

At first glance, `nonlocal` doesn't seem to change the lookup rule in any significant way -- the search would still continue from the next scope down the stack. However, one subtle behavior here is that `nonlocal` limits the search to *local scopes* only. Global scope is ignored. For example,

```
x = 1
def foo():
  def bar():
    nonlocal x
    x    # `x` should not be considered a bound variable here, despite the existence of a global binding with the same name
```

 ---

The return type of `ScopeStack.lookup` is worth mentioning. In addition to the binding itself, I choose to also include (1) the scope where the binding is found, and (2) some metadata about the said scope and the nature of this lookup. Those metadata include whether the said scope is on the top of the scope stack (i.e. `CurrentScope` vs. `OuterScope`), as well as whether the looked up name is declared as `nonlocal` or `global` in the current scope (i.e. `Modifier.Local` vs. `Modifier.Nonlocal` vs. `Modifier.Global`). The info can be useful to determine what bindings to inject to the AST that would be useful to the typechecker -- (1) bindings that do not come from current scope (as function-level checking can easily find them, and (2) bindings that do not come from module-level scope (as they'll be included in the global environment).

Reviewed By: dkgi

Differential Revision: D18441825

fbshipit-source-id: 75777eeb7790472bd2a692c206a808f4e389f4fc