Fix noop call assertions in upgrade test
Summary:
howdidthishappen

Lots of these tests were never getitng run. This was hiding issues with iterable objects being unequal to one another even if they were created on the same input, so need to mock out `errors.sort_errors`.

TIL that when mocking in python, you mock out the function as qualified by the module importing it, not the module defining it...

Reviewed By: pradeep90

Differential Revision: D18510030

fbshipit-source-id: e69042e65a89f970f5a647689747cd2855cd81ef