Moving check_class_hierarchy_integrity into ClassHierarchyEnvironment.ReadOnly
Summary:
I'd like to change things up such that you can only access the read only interface of an environment after updating it.  There are only two places that kind of do that now, and this diff deals with the second one: running GlobalResolution.check_class_hierarchy_integrity before we make the AnnotatedGlobalEnvironment that defines a global resolution.

This is just an easy move.

Reviewed By: sinancepel

Differential Revision: D18522015

fbshipit-source-id: 43081c84a82bdc609e321409c7fedd2a0007c76f