Making a version of IncrementalCheck.recheck that does not need a full server state
Summary:
I'd like to change things up such that you can only access the read only interface of an environment after updating it.  There are only two places that kind of do that now, and this diff deals with one: SavedState, who makes an environment before actually updating it with IncrementalCheck.recheck.
We can avoid this problem by breaking up the state and only passing the elements of it that IncrementalCheck.recheck actually uses, which can't include environment anyway, since it's actively updating it.

Reviewed By: sinancepel

Differential Revision: D18521048

fbshipit-source-id: d3763a829fc8f99ddce2a185a73ec56ca9c37711