Defining legacy_invalidated_keys in terms of the unannotated_global_environment
Summary:
I think I had planned to do this earlier as you can see some traces of this in the UnannotatedGlobalEnvironment.ReadOnly.

The idea is that all of the environments define their legacy key sets in terms of the unannotateglobalenvironment, so we can reduce the amount of "tunnelling" we need to do through layers of environments by always giving them what they want.  This drives towards  my objective of only needing the immediate successor to change when adding a new environment.

Reviewed By: grievejia

Differential Revision: D18522985

fbshipit-source-id: 05a9a9c5c0852b229b90612410d87fca6f9dee98