Removing the need to "create" Environments
Summary:
In the transition to managed caches, we're going to end up creating a bunch more environments.  As of now, that requires adding tons of annoying boilerplate in clients each time to add a new call to `create` and `read_only` and threading it through.

Ultimately, those interfaces are counterproductive for what we'd really like to ensure: that we're not accessing environments before we've updated/populated them.  I was able to enforce that with a cool maneuver, having update itself return the read-only interface.  This enforces the property discussed above, which previously was only weakly enforced with easy ways to break it, and greatly reduces the amount of client boilerplate, by only needing new environments to be added to the update chain, which is unavoidable.

Reviewed By: grievejia

Differential Revision: D18512537

fbshipit-source-id: 9e286b516f7ae6b8d993db06c5787c202d843a69